fix: apply review findings and record the declined one
Stale scene count in the combinatorics README -- the render list and the
'all four' comment predate WhenToUseIt. The 'four elementary counting rules'
line in Scope is left alone: there are still four rules, and the fifth scene
is a synthesis, not a fifth rule.

Dropped a redundant int() in rgb_to_hex. I expected removing it to break the
f-string, since a numpy scalar has no 'x' format -- but round(np.float64)
returns a plain int here, verified before changing it.

Declined again, and now recorded in docs/decisions.md so it stops recurring:
re-exporting utils.colour through the package root. The root is what concept
modules import from, and a scene never computes a colour distance -- only the
solver and the tests do, and both import the submodule directly. This is the
second time it has been raised, which is the exact failure the decisions file
was added for, and I did not write the entry when I declined it the first
time.

+++ utils/colour.py
@@ -66,7 +66,7 @@ def rgb_to_hex(rgb: np.ndarray) -> str:
     measuring distances between unshowable values is what caused the bug in
     the docstring above.
     """
-    red, green, blue = (int(round(v * 255)) for v in np.clip(rgb, 0, 1))
+    red, green, blue = (round(v * 255) for v in np.clip(rgb, 0, 1))
     return f"#{red:02x}{green:02x}{blue:02x}"
